Пофикшены страницы после ревью
