+++ src/utils/utils.ts
@@ -19,14 +19,6 @@ export const swapArrElements = (
   [arr[firstIndex], arr[secondIndex]] = [arr[secondIndex], arr[firstIndex]];
 };
 
-export const getFibanacciArr = (n: number): number[] => {
-  let resultArr: number[] = [0, 1];
-  for (let i = 2; i <= n + 1; i++) {
-    resultArr.push(resultArr[i - 2] + resultArr[i - 1]);
-  }
-  return resultArr.slice(1);
-};
-
 export const getRandomArr = (
   minArrLength: number,
   maxArrLength: number,
